+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add `AmazonJWT` module
 
 ## 5.0.0 - 2024-06-05
 - Add `Action` type with 2 cases